Remove wrapper method from OlmMachine
The store object is accessible, so there is no need to provide a wrapper in
`OlmMachine`.
